extract book routes to route book
